Duplicated Taskfile.yml
Problem:
There is duplicated logic in 2 Taskfiles - one
for inside the container and one for outside.

Solution:
Test the $USER variable for accepted values from a vscode or
builder container and run cmds without 'docker run' prefixed
o found.
Additionally pre-install nodejs for the pyright type-checker so
that pyright checks complete quickly.

+++ scripts/notebooks.sh
@@ -2,6 +2,13 @@
 #
 # run jupyter notebooks
 #
+
+if [ "$USER" = "builder" -o "$USER" = "vscode" ]
+then
+    echo "Cannot run notebooks inside container"
+    exit 0
+fi
+
 docker run --rm -it \
     -v $(pwd):/home/builder \
     -u $(id -u):$(id -g) \
